requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6